// src/generated/olympusUnits.ts
export type EpochsEpoch = {
  createdAt: string;
  endTimestamp: number;
  epochNumber: number;
  id: number;
  rewardStatuses: string[];
  seasonId: number;
  seasonName: string;
  startTimestamp: number;
  status: string;
  updatedAt: string;
};

export type SharedEpochRewardsStatus =
  (typeof SharedEpochRewardsStatus)[keyof typeof SharedEpochRewardsStatus];

export const SharedEpochRewardsStatus = {
  pending: "pending",
  calculated: "calculated",
  distributed: "distributed",
} as const;

export type EpochsEpochReward = {
  createdAt: string;
  epochId: number;
  id: number;
  merkleRoot: string | boolean | null;
  rewardAmount: string;
  rewardAssetId: number;
  safeTxHash: string | boolean | null;
  safeUrl: string | boolean | null;
  status: SharedEpochRewardsStatus;
  tokenAddress: string;
  tokenDecimals: number;
  tokenSymbol: string;
  updatedAt: string;
};